[portenta] Send service pack updates through mbed initialization
The mbed ble stack is initialized at the beginning in order to exploit the service pack transfer.
Once it has been initialized, the queue used to dispatch the ble events is stopped.
At this point the mbed stack is no more used and ArduinoBLE exploits directly the low level uart transport layer.