Perform remaining length checks on advertising parameters set methods
Now all the set methods of advertising data parameters return a bool value.
This bool indicates whether the parameter has been correctly set or not, depending on the size of the passed value.
A remaining length parameter of advertising data has been introduced. It can be publicly read, it indicates how many bytes can be still written into the packet.