Change allocation of default services and characteristics
Default local characteristics and services are now allocated in the HEAP.
This change is needed for CI integration. Indeed, if these attributes are allocated in the STACK they cannot be freed by the 'delete' statement used in their destructors.
Downside: their memory usage will not be counted in the estimated amount of RAM occupated by a resulting program that includes this library.