Change advertising raw data setting
If a BLEAdvertisingData packet has the raw data parameter set, all the other parameters are ignored when encoding the actual data packet.
This means that the raw data parameter can be as long as the maximum length for an advertising packet (MAX_AD_DATA_LENGTH), that is 31 bytes as per BLE standard.
During the setting of the raw data parameter: if the passed raw data length is larger than this maximum value, then the raw data is not set and the function returns false.