Fix broken image embed in documentation
The file previously targeted by the embed has been lost or moved. Experience shows that this is is a common occurrence
with the files hosted by Arduino, and will be especially problematic in cases like this where the file may appear to be
superfluous to the website maintainers due to the lack of any internal references to it, so the only way to ensure the
file will continue to be available is by hosting alongside the documentation content in the repository.

I was able to recover the lost file from the Internet Archive Wayback Machine's 2023-06-18 archive of the previous URL:

https://web.archive.org/web/20230618230604/https://www.arduino.cc/en/uploads/Reference/ble-bulletin-board-model.png

The absolute URL of the file was intentionally used in the markup instead of what would normally be the best practices
of using the relative path. The reason is because the documentation content will be published at
https://reference.arduino.cc/reference/en/libraries/arduinoble/ and the publishing system doesn't have any capability
for handling asset files.