Fix broken link in documentation
The previous target of this link is no longer available.

I found that there is no longer an authoritative web page that provides a list of standard GATT service UUIDs, equivalent to the previous target of that link:

https://web.archive.org/web/20160305021019/https://developer.bluetooth.org/gatt/services/Pages/ServicesHome.aspx

The Bluetooth Special Interest Group has since moved this information into the "Assigned Numbers" pdf, so I replaced the
broken link with a link to the download page for that document and noted its location in that document.