Add BLEPermission Support to BLEStringCharacteristic
5 years ago, support for encryption was added by #156.
Currently, the BLEEncryption tag needs to be added to require proper encrypted bluetooth pairing.
It is possible to add this tag to all other kinds of characteristics except String characteristics.

BLEStringCharacteristic's constructor calls BLECharacteristic's constructor, and BLECharacteristic
uses a uint16_t instead of a uint8_t.

This PR simply changes the properties attribute of BLEStringCharacteristic to match the width of
the permissions attribute of BLECharacteristic (from 8 to 16 bits) to support the BLEPermission
tags in BLEProperty.h.