Modernize spell checking infrastructure
The repository contains a CI workflow to check for commonly misspelled words in the project files. This is done using
the "codespell" tool.

A GitHub Actions action handles the installation and execution of the tool. At the time Arduino started setting up such
infrastructure in project repositories, no such action was available and so Arduino created an action. Since that time,
the codespell developers created an equivalent action and so Arduino's action was abandoned. Most of the libraries were
migrated to the new action, but somehow this repository is still using the abandoned action. Due to bit rot, Arduino's
action now fails to initialize:

```
 #8 [3/4] RUN pip3 install codespell
 #8 0.401 error: externally-managed-environment

[...]

ERROR: failed to solve: process "/bin/sh -c pip3 install codespell" did not complete successfully: exit code: 1
Error: Docker build failed with exit code 1
```

The problem is solved by migrating the spell check workflow to using the codespell action.