Fix typos
The "codespell" spellchecker tool is used to automatically detect commonly misspelled words in the files of this
project.

The misspelled words dictionary was expanded in recent releases of codespell, which resulted in the detection of
misspelled words in the project files:

> Error: ./src/utility/HCI.cpp:761: Signalling ==> Signaling
> Error: ./docs/readme.md:22: acknowledgement ==> acknowledgment

The typos are hereby corrected, which will restore the spell check to a passing state.

---

I supplemented the correction of the automatically detected typos with a general review of the project content,
fixing the additional typos I identified during that review.