EncryptedBatteryMonitor: fix memory management bugs
There was a buffer overflow in the EncryptedBatteryMonitor example code
when formatting the 6-digit confirmation code for display during
pairing, since the terminating null character was not accounted for in
the buffer size.

Also, there was a memory leak due to unnecessary dynamic allocation
when setting the initial value of the string characteristic.

Signed-off-by: Luca Burelli <l.burelli@arduino.cc>